Add files via upload
Here are the original source codes, however I have converted them from .ipynb to .py. That way these can be ran without needing an jupyter notebook enviornment. 

The third file, QCiM.py is a combo of the other two (QualityControl.py and PyWindowQCiM.py). The order needs to be reversed, but gives good endorsement that the project is headed in the right direction.